requirements.txt
blas==1.0=mkl
configparser==4.0.2
cudatoolkit==10.1.243
hdf5==1.10.2
intel-openmp==2022.0.1
ipykernel==4.10.0
ipython_genutils==0.2.0
ipywidgets==7.6.0
libopencv==3.4.2
mkl==2020.2
mkl-service==2.3.0
mkl_fft==1.0.15
mkl_random==1.1.0
ninja==1.10.2
numpy==1.16.6
numpy-base==1.16.6
opencv==3.4.2
pillow==6.2.1
pip==19.3.1
py-opencv==3.4.2
pycparser==2.20
pyparsing==2.4.7
python==2.7.18
python-dateutil==2.8.2
pytorch==1.4.0+cu101
six==1.16.0
torchaudio==0.4.0
torchvision==0.5.0
wheel==0.37.1
widgetsnbextension==3.5.1
zlib==1.2.11
ipdb==0.13.9
ipython==5.10.0
kiwisolver==1.1.0
matplotlib==2.2.5
pickleshare==0.7.5
progress==1.6
protobuf==3.17.3
pyyaml==5.4.1
scikit-learn==0.20.4
scipy==1.2.3
sklearn==0.0
subprocess32==3.5.4
tqdm==4.63.0
